MSNP11 porting:
I ported the contact and group handling. That means the adding, moving, removing groups and contact.
Rewritten the LST, LSG, ADG, REG, REM.
Changed the signal and slot related to these functions.

Added ADC command and deprecated ADD command.


svn path=/branches/work/kopete/msnp11_porting/; revision=433130

+++ protocols/msn/msnprotocol.cpp
@@ -91,7 +91,7 @@ Kopete::Contact *MSNProtocol::deserializeContact( Kopete::MetaContact *metaConta
 	MSNContact *c = new MSNContact( account, contactId, metaContact );
 
 	for( QStringList::Iterator it = groups.begin() ; it != groups.end(); ++it )
-		c->contactAddedToGroup( ( *it ).toUInt(), 0L  /* FIXME - m_groupList[ ( *it ).toUInt() ]*/ );
+		c->contactAddedToGroup( *it, 0L  /* FIXME - m_groupList[ ( *it ).toUInt() ]*/ );
 
 	c->m_obj= serializedData[ "obj" ];
 	c->setInfo( "PHH" , serializedData[ "PHH" ] );
